refactor: split git_skill_service.py
